refactor: extract RecentDomainsSection into reusable component (#41020)
### Description

Refactored the recent domains section from being an inline constant in
`SignUp.tsx` into a dedicated, reusable React component. This improves
code organization, maintainability, and testability.

### Changes Made

- **Created new component**: `RecentDomainsSection.tsx`
  - Encapsulates all logic for rendering recent organization domains
- Handles conditional rendering internally (returns null when no
domains)
  - Includes proper TypeScript types and imports

- **Updated SignUp component**:
  - Removed inline `recentDomainsSection` constant definition
  - Added import for new `RecentDomainsSection` component
  - Replaced usage with JSX component

- **Improved avatar logic**:
- Changed from alphabetical cycling (`String.fromCharCode(65 + (index %
26))`) to using the actual organization name's first letter
(`orgName.charAt(0).toUpperCase()`)
  - This provides more meaningful and recognizable avatars for users

## Automation

/ok-to-test tags="@tag.Sanity, @tag.Authentication"

### :mag: Cypress test results
